% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_trials}
\alias{split_trials}
\title{Split a vector of trials into fixed-size blocks}
\usage{
split_trials(x, k)
}
\arguments{
\item{x}{A vector. Usually a vector of trial indices or observations.}

\item{k}{Integer. The number of elements per block.}
}
\value{
An integer vector of the same length as \code{x}, giving the block number
for each element.
}
\description{
Assigns each element of a vector to a block of size \code{k}. The last block may
contain fewer than \code{k} elements if the length of \code{x} is not an exact multiple
of \code{k}.
}
\examples{
split_trials(1:10, k = 3)

}
