% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{block2trial}
\alias{block2trial}
\title{Convert block indices to cumulative trial numbers}
\usage{
block2trial(x, start = 32, step = 5)
}
\arguments{
\item{x}{A vector of block indices. Only the length of \code{x} is used.}

\item{start}{Integer. Number of trials in the first block. Default is \code{32}.}

\item{step}{Integer. Number of additional trials added after the first block.
Default is \code{5}.}
}
\value{
An integer vector of cumulative trial counts.
}
\description{
Converts a sequence of block indices into cumulative trial counts given an
initial block size and a fixed step size for subsequent blocks.
}
\examples{
block2trial(1:5, start = 32, step = 5)

}
